user friendly install script
Created this script to help my coworkers remotely install certain programs from their local machine to the remote one. Most of them do not use powershell so I had to make a script that is more user friendly and they dont need to change the paths for this to work. It also installs PSEXEC so they can remotely turn on quickrm for the script to work.