Helps declutter computer
We have some old machines with too many profiles and that slows down the computer and leaves no internal storage. This is still in testing as there are only a few of those old machines. It has worked but I am not fully confident it will ALWAYS work as expected since I always go in and make some changes